+++ src/types/bindings/Page.ts
@@ -3,6 +3,7 @@ import type { Block } from "./Block";
 import type { EditorPreferences } from "./EditorPreferences";
 import type { PageAnnotations } from "./PageAnnotations";
 import type { PageId } from "./PageId";
+import type { PageProtection } from "./PageProtection";
 import type { SectionId } from "./SectionId";
 
-export type Page = { id: PageId, section_id: SectionId, title: string, tags: Array<string>, blocks: Array<Block>, annotations: PageAnnotations, editor_preferences: EditorPreferences, pdf_asset: string | null, pdf_total_pages: number | null, canvas_state: any | null, created_at: string, updated_at: string, schema_version: number, };
+export type Page = { id: PageId, section_id: SectionId, title: string, tags: Array<string>, blocks: Array<Block>, annotations: PageAnnotations, editor_preferences: EditorPreferences, pdf_asset: string | null, pdf_total_pages: number | null, canvas_state: any | null, created_at: string, updated_at: string, schema_version: number, protection: PageProtection | null, encrypted_content: string | null, };